+++ src/group/group.router.ts
@@ -1,16 +1,15 @@
-import { createFeatureRouter } from './endpoints/group.endpoint';
-
+import { createFeatureRouter } from '../utils/endpoint';
+import CreateGroup from './endpoints/group.create';
+import DeleteGroup from './endpoints/group.delete';
 import GetGroupByID from './endpoints/group.getByID';
 import GetGroupByUserID from './endpoints/group.getByUser';
 import SearchGroupByName from './endpoints/group.searchByName';
-import CreateGroup from './endpoints/group.create';
 import UpdateGroup from './endpoints/group.update';
-import DeleteGroup from './endpoints/group.delete';
 
-import GetUsersOfGroup from './user/enpoints/group.user.get';
-import AddUserToGroup from './user/enpoints/group.user.add';
-import UpdateUserRole from './user/enpoints/group.user.update';
-import RemoveUserFromGroup from './user/enpoints/group.user.remove';
+import AddUserToGroup from './user/endpoints/group.user.add';
+import GetUsersOfGroup from './user/endpoints/group.user.get';
+import RemoveUserFromGroup from './user/endpoints/group.user.remove';
+import UpdateUserRole from './user/endpoints/group.user.update';
 
 export default createFeatureRouter(
   new GetGroupByID(),         // GET    /:id
